change data type of field point - int to long

+++ lms/lmsbackend/coderunnerservice/src/main/java/com/hust/baseweb/applications/programmingcontest/entity/ContestSubmissionEntity.java
@@ -64,7 +64,7 @@ public class ContestSubmissionEntity {
     private float memoryUsage;
 
     @Column(name = "point")
-    private Integer point;
+    private Long point;
 
     @Column(name = "status")
     private String status;
